Updated the create ISO script to add a few linux boot files from ISOLINUX

+++ create-iso.sh
@@ -4,36 +4,54 @@
 #  It will be a "hybrid" image that you can put onto USB using disk-damager
 #
 echo "Creating development directories in ./binary"
-mkdir -p binary/{live,isolinux}
-echo "Copying LINUX to the development directory"
+rm -rf binary # remove old dev files
+mkdir -p binary/{live,isolinux} # create a few directories
+echo "Copying LINUX to the development directory" # copy over LINUX
 cp chroot/boot/vmlinuz-4.4.0-1-686-pae binary/live/vmlinuz
 cp chroot/boot/initrd.img-4.4.0-1-686-pae binary/live/initrd
 echo "Making SquashFS filesystem, this will take some time."
 mksquashfs chroot binary/live/filesystem.squashfs -comp xz -e boot
-cp /usr/lib/ISOLINUX/isolinux.bin binary/isolinux/
-cp /usr/lib/syslinux/modules/bios/menu.c32 binary/isolinux/
-# Create an ISOLinux menu:
-echo "Creating the ISOLINUX configuration file" # this can be scripted to take args later:
-cat > binary/isolinux/isolinux.cfg << CFG
-ui menu.c32
-prompt 0
-menu title Weakerthan Linux 7 Elite
-timeout 300
 
-label wt7-elite-live
- menu label ^Weakerthan Linux 7 Elite - (Live)
- menu default
- linux live/vmlinuz
- append initrd=/live/initrd boot=live persistence quiet
-CFG
-# Create the ISO with timestamp:
+# CREATE ISO
 echo "Creating the ISO image, this will take some time."
-TS=$(date|awk '{gsub(":",".",$4); print $3$4$6}')
-xorriso -as mkisofs -r -J -joliet-long -l -cache-inodes \
- -isohybrid-mbr /usr/lib/syslinux/isohdpfx.bin -partition_offset 16 \
- -A "Debian Live"  -b isolinux/isolinux.bin -c \
- isolinux/boot.cat -no-emul-boot -boot-load-size 4 \
- -boot-info-table -o wt7-elite-$TS.iso binary
-# Create an md5 of the ISO:
-echo "Craeting MD5 checksum for your ISO image."
-md5sum wt7-elite-$TS.iso |awk '{print $1}' > wt7-elite-$TS.md5
+if [ -f "binary/live/filesystem.squashfs" ];then # filesystem successfully made
+ cp /usr/lib/ISOLINUX/isolinux.bin binary/isolinux/ # ISOLINUX for booting into live ISO
+ cp /usr/lib/syslinux/modules/bios/menu.c32 binary/isolinux/
+ cp /usr/lib/syslinux/modules/bios/hdt.c32 binary/isolinux/
+ cp /usr/lib/syslinux/modules/bios/ldlinux.c32 binary/isolinux/
+ cp /usr/lib/syslinux/modules/bios/libcom32.c32 binary/isolinux/
+ cp /usr/lib/syslinux/modules/bios/libutil.c32 binary/isolinux/
+ # Create an ISOLinux menu:
+ echo "Creating the ISOLINUX configuration file" # this can be scripted to take args later:
+ cat > binary/isolinux/isolinux.cfg << CFG
+ ui menu.c32
+ prompt 0
+ menu title Weakerthan Linux 7 Elite
+ timeout 300
+ 
+ label wt7-elite-live
+  menu label ^Weakerthan Linux 7 Elite - (Live)
+  menu default
+  linux live/vmlinuz
+  append initrd=/live/initrd boot=live persistence quiet
+CFG
+ TS=$(date|awk '{gsub(":",".",$4); print $3$4$6}')
+ xorriso -as mkisofs -r -J -joliet-long -l -cache-inodes \
+  -isohybrid-mbr /usr/lib/ISOLINUX/isohdpfx.bin -partition_offset 16 \
+  -A "Debian Live"  -b isolinux/isolinux.bin -c \
+  isolinux/boot.cat -no-emul-boot -boot-load-size 4 \
+  -boot-info-table -o wt7-elite-$TS.iso binary
+else
+ echo "SquashFS not created. Something went wrong. Please see details above"
+ exit 1;
+fi 
+
+# CREATE MD5
+echo "Creating MD5 checksum for your ISO image."
+if [ -f "wt7-elite-$TS.iso" ]; then # file exists:
+ echo "ISO file created successfully, generating MD5"
+ md5sum wt7-elite-$TS.iso |awk '{print $1}' > wt7-elite-$TS.md5
+else # no ISO
+ echo "ISO was not created, something went wrong. Please see details above."
+ exit 1;
+fi
